differs copy color by click position; fixes #176
Instead of using the Alt key (which caused problems depending on the
operating system), now it copies either the color value or the source
(variable/function) by clicking either of it. By default, the value is
copied unless the curser points at the source. Which one will be copied
is indicated by underlining it. Additionally the copy icon is displayed
within the color swatch on hover.

As a side effect this avoids the tooltip which was English, what might
be a problem for users of other languages.